Added the basic Verilog code structure
I think the glory of matlab is over, now time to test and see if we can get this working. The compare to threshold code it disgusting and needs to be rewritten but otherwise it is on the right track. Still need to find out how to implement the low pass filter